Added serialization technique to Stateful and Stateless service extension methods that attempts an optional ExceptionSerialization. This defaults to BinaryFormatter so as not to break existing implementations.
Added another overload that allows the user to provide an aray of IExceptionConvertors. If this is provided, the user intends to use the DSC serialization.

Marked the implementing extension for both Stateful and Stateless services internal so as not to confuse users with the ability to both pass an array of IExceptionConvertor while still setting the serialization technique to BinaryFormatter.