Avoid locking on each call to ServicePartitionResolver.GetDefault()
Each call to ServicePartitionResolver.GetDefault() acquires a lock due
to lazy initialization. We could use double-checked locking to avoid
this, but as the object is not heavy and its creation is not expensive
we can simply do eager initialization.

ServicePartitionResolver.SetDefault() can update this field, but reads
and writes of reference types are atomic. If needed, field can be
declared as volatile.