Deprecating Service Remoting V1 stack
This is the first step towards moving to Service Remoting V2 completely.

All code that is part of the Service Remoting V1 stack have been marked as obsolete and will be removed in the future.
In parts of the code that reference code form both the V1 and V2 stack, pragmas are used to suppress CS0618 (obsolete) warnings, so that the SDK packages can be built. Once the transition to V2 is fully complete, this suppression will not be needed.

From now on, when using service or actor remoting, it is necessary to specify the remoting stack explicitly using assembly attributes, otherwise an exception will be raised. User who are already using V2 stack will not see a difference after upgrading to the new SDK. Users who are still using V1 stack by default will have to make a conscious decision to continue using V1 stack or to upgrade to V2 when upgrading to the new SDK.
# Conflicts:
#	properties/service_fabric_common.props