Remove not needed double projects (#408)
We used to have multiple projects that would essentially just serve to
build the same code in NET Framework and NET Standard respectively.
Since we use SDK style projects, there is no need for this, since they
allow 1 project to be built with multiple frameworks.
This PR removes old projects, and updates configuration for the original
projects to be able to be built using NET Standard.