Remove Services.Tests.csproj from list of test projects in netframework.targets
This project doesn't target net472 and fails to build in the ADO pipeline due to
a missing Main entry point. It is not clear why this error doesn't happen during
GitHub builds, however since we're removing the old build infrastructure, it's
probably not worth the effort to get to the root cause of this.