Make Microsoft.ServiceFabric.Actors package final FabActUtil
The Pack target appears to have additional logic to transform the BuildOutputInPackage
items from their intermediate to the final output paths. However, for the
TfmSepecificPackageFile items, it uses the original item specification.

The previous version of the PackageContent target used final version of
the Microsoft.ServiceFabric.Actors assembly and the intermediate version
of the FabActUtil. Since the the ADO build pipeline will be signing the
final build outputs only, this would leave the FabActUtil.exe/.dll unsigned.

The new version of the PackageContent target uses the publicly-documented
BuiltProjectOutputGroup target instead of the internal _GetBuildOutputFilesWithTfm.
Both targets return the intermediate project outputs, which have to be
transformed to their final output paths explicitly.