Add FabricTransport.Tests source code
This code doesn't build yet because we use the public
signing key to partially sign test assemblies in this
repo, however the System.Fabric.dll makes InternalsVisibleTo
the assembly signed with the internal test key.

Once the System.Fabric.dll is updated to make InternalsVisibleTo
the FabricTransport.Tests.dll signed with the public key
we will include the project in the solution and enable tests.